requirements.txt
dotenv
boto3
django-storages==1.14